pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genomic_lightning"
version = "0.1.0"
description = "PyTorch Lightning framework for genomic deep learning models"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pytorch-lightning>=2.0.6",
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "torchmetrics>=0.11.4",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "h5py>=3.7.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0", 
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.9.0",
]
logging = [
    "wandb>=0.12.0",
    "tensorboard>=2.8.0",
]
genomics = [
    "pyfaidx>=0.7.0",
    "pysam>=0.19.0",
    "captum>=0.6.0",
    "biopython>=1.80",
]

[project.scripts]
genomic_lightning = "genomic_lightning.cli.main:cli"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
